requirements.txt
uuid
hashlib
secrets
cryptography
dnspython